Fix duplicated code
