+++ src/main/java/org/folio/marccat/dao/persistence/AuthorityNote.java
@@ -1,29 +1,20 @@
 package org.folio.marccat.dao.persistence;
 
-import org.folio.marccat.business.cataloguing.bibliographic.VariableField;
-import org.folio.marccat.business.cataloguing.common.OrderedTag;
-import org.folio.marccat.business.common.PersistenceState;
-import org.folio.marccat.business.common.PersistentObjectWithView;
 import org.folio.marccat.business.common.View;
-import org.folio.marccat.config.constants.Global;
-import org.folio.marccat.dao.AbstractDAO;
-import org.folio.marccat.shared.CorrelationValues;
-import org.folio.marccat.util.StringText;
 
 /**
  * @author elena
  *
  */
-public class AuthorityNote extends VariableField implements PersistentObjectWithView, OrderedTag {
-	private static final short NOTE_CATEGORY = 7;
-	private String content = null;
-	private int noteType;
-	private int noteNumber = -1;
-	private Integer sequenceNumber;
+public class AuthorityNote extends RecordNote {
+	
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 5041286137447344231L;
 
 	public AuthorityNote() {
 		super();
-		setPersistenceState(new PersistenceState());
 	}
 
 	/**
@@ -32,38 +23,10 @@ public class AuthorityNote extends VariableField implements PersistentObjectWith
 	public AuthorityNote(final int itemNbr) {
 		super(itemNbr);
 	}
-
+	
 	@Override
-	public boolean isBrowsable() {
-		return false;
-	}
-
-	/**
-	 *
-	 */
-	public int getNoteNumber() {
-		return noteNumber;
-	}
-
-	/**
-	 * @param i
-	 */
-	public void setNoteNumber(final int i) {
-		noteNumber = i;
-	}
-
-	/**
-	 *
-	 */
-	public String getStringTextString() {
-		return content;
-	}
-
-	/**
-	 * This value is used by Hibernate It must preserve the $a value
-	 */
-	public void setStringTextString(final String st) {
-		this.content = st;
+	public int hashCode() {
+		return getNoteNbr();
 	}
 
 	/*
@@ -77,71 +40,8 @@ public class AuthorityNote extends VariableField implements PersistentObjectWith
 			return false;
 		} else {
 			AuthorityNote aNote = (AuthorityNote) obj;
-			return aNote.getAutItemNumber() == this.getAutItemNumber() && aNote.getNoteNumber() == this.getNoteNumber();
+			return aNote.getAutItemNumber() == this.getAutItemNumber() && aNote.getNoteNbr() == this.getNoteNbr();
 		}
-	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see java.lang.Object#hashCode()
-	 */
-	@Override
-	public int hashCode() {
-		return getNoteNumber();
-	}
-
-	public StringText getStringText() {
-		return new StringText(content);
-	}
-
-	/**
-	 * @param text should not ends with Subfield separator
-	 */
-	public void setStringText(final StringText text) {
-		content = text.toString();
-	}
-
-	public int getCategory() {
-		return NOTE_CATEGORY;
-	}
-
-	@Override
-	public CorrelationValues getCorrelationValues() {
-		return null;
-	}
-
-	public void setCorrelationValues(final CorrelationValues v) {
-		setNoteType(v.getValue(1));
-	}
-
-	/**
-	 *
-	 */
-	public int getNoteType() {
-		return noteType;
-	}
-
-	/**
-	 *
-	 */
-	public void setNoteType(final int s) {
-		noteType = s;
-	}
-
-	@Override
-	public AbstractDAO getDAO() {
-		return null;
-	}
-
-	@Override
-	public boolean correlationChangeAffectsKey(CorrelationValues v) {
-		if (v.getValue(1) == Global.PUBLISHER_DEFAULT_NOTE_TYPE) {
-			return true;
-		} else if (v.getValue(1) == 381) {
-			return true;
-		} else
-			return v.getValue(1) == 382;
 	}
 
 	public int getAutItemNumber() {
@@ -150,18 +50,6 @@ public class AuthorityNote extends VariableField implements PersistentObjectWith
 
 	public void setAutItemNumber(int i) {
 		setItemNumber(i);
-	}
-
-	public void setContent(String text) {
-		this.content = text;
-	}
-
-	public Integer getSequenceNumber() {
-		return sequenceNumber;
-	}
-
-	public void setSequenceNumber(Integer integer) {
-		sequenceNumber = integer;
 	}
 
 	@Override
