+++ src/main/java/org/folio/marccat/dao/persistence/BibliographicNote.java
@@ -4,16 +4,9 @@
  */
 package org.folio.marccat.dao.persistence;
 
-import org.folio.marccat.business.cataloguing.bibliographic.VariableField;
-import org.folio.marccat.business.cataloguing.common.OrderedTag;
-import org.folio.marccat.business.common.PersistenceState;
-import org.folio.marccat.business.common.PersistentObjectWithView;
 import org.folio.marccat.business.common.UserViewHelper;
-import org.folio.marccat.config.constants.Global;
 import org.folio.marccat.dao.AbstractDAO;
 import org.folio.marccat.dao.BibliographicNoteDAO;
-import org.folio.marccat.shared.CorrelationValues;
-import org.folio.marccat.util.StringText;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,22 +15,16 @@ import java.util.List;
  *
  * @author janick
  */
-public class BibliographicNote extends VariableField implements  PersistentObjectWithView, OrderedTag {
-  private static final short NOTE_CATEGORY= 7;
+public class BibliographicNote extends RecordNote {
   private List<BibliographicNoteOverflow> overflowList = new ArrayList<>();
   /**
    * The content is all or a part of the stringText.toString() result.
    */
-  private String content = null;
-  private int noteType;
-  private int noteNbr = -1;
   private char overflowIndicator = '0';
   private UserViewHelper userViewHelper = new UserViewHelper();
-  private Integer sequenceNumber;
 
   public BibliographicNote() {
     super();
-    setPersistenceState(new PersistenceState());
   }
 
   /**
@@ -46,40 +33,11 @@ public class BibliographicNote extends VariableField implements  PersistentObjec
   public BibliographicNote(final int itemNbr) {
     super(itemNbr);
   }
-
+  
   @Override
-  public boolean isBrowsable() {
-    return false;
-  }
-
-  /**
-   *
-   */
-  public int getNoteNbr() {
-    return noteNbr;
-  }
-
-  /**
-   * @param i
-   */
-  public void setNoteNbr(final int i) {
-    noteNbr = i;
-  }
-
-  /**
-   *
-   */
-  public String getStringTextString() {
-    return content;
-  }
-
-  /**
-   * This value is used by Hibernate
-   * It must preserve the $a value
-   */
-  public void setStringTextString(final String st) {
-    this.content = st;
-  }
+	public int hashCode() {
+		return getNoteNbr();
+	}
 
   /* (non-Javadoc)
    * @see java.lang.Object#equals(java.lang.Object)
@@ -94,52 +52,6 @@ public class BibliographicNote extends VariableField implements  PersistentObjec
         aNote.getUserViewString().equals(this.getUserViewString()) &&
         aNote.getNoteNbr() == this.getNoteNbr();
     }
-  }
-
-  /* (non-Javadoc)
-   * @see java.lang.Object#hashCode()
-   */
-  @Override
-  public int hashCode() {
-    return getNoteNbr();
-  }
-
-  public StringText getStringText() {
-    return new StringText(content);
-  }
-
-  /**
-   * @param text should not ends with Subfield separator
-   */
-  public void setStringText(final StringText text) {
-    content = text.toString();
-  }
-
-  public int getCategory() {
-    return NOTE_CATEGORY;
-  }
-
-  @Override
-  public CorrelationValues getCorrelationValues() {
-    return null;
-  }
-
-  public void setCorrelationValues(final CorrelationValues v) {
-    setNoteType(v.getValue(1));
-  }
-
-  /**
-   *
-   */
-  public int getNoteType() {
-    return noteType;
-  }
-
-  /**
-   *
-   */
-  public void setNoteType(final int s) {
-    noteType = s;
   }
 
   /**
@@ -168,7 +80,6 @@ public class BibliographicNote extends VariableField implements  PersistentObjec
     return overflowList;
   }
 
-
   /**
    * @param list
    */
@@ -176,47 +87,22 @@ public class BibliographicNote extends VariableField implements  PersistentObjec
     overflowList = list;
   }
 
-
   @Override
-  public boolean correlationChangeAffectsKey(CorrelationValues v) {
-    if (v.getValue(1) == Global.PUBLISHER_DEFAULT_NOTE_TYPE) {
-      return true;
-    } else if (v.getValue(1) == 381) {
-      return true;
-    } else return v.getValue(1) == 382;
-  }
-
-
   public String getUserViewString() {
     return userViewHelper.getUserViewString();
   }
 
-
+  @Override
   public void setUserViewString(String string) {
     userViewHelper.setUserViewString(string);
   }
-
 
   public int getBibItemNumber() {
     return getItemNumber();
   }
 
-
   public void setBibItemNumber(int i) {
     setItemNumber(i);
   }
-
-  public void setContent(String text) {
-    this.content = text;
-  }
-
-  public Integer getSequenceNumber() {
-    return sequenceNumber;
-  }
-
-  public void setSequenceNumber(Integer integer) {
-    sequenceNumber = integer;
-  }
-
 
 }
